feat: restructure into cronalytics/ package namespace
Move all flat .py modules into cronalytics/ directory to enable
proper pip install with namespace isolation:

- Moved: cli.py, config.py, facts.py, ingester.py, logger.py,
  scanner.py, schedule.py -> cronalytics/
- Added: cronalytics/__init__.py
- Updated root __init__.py for absolute imports
- Updated dashboard/plugin_api.py _load_module() path
- Updated cronalytics/cli.py bootstrap and imports
- Updated cronalytics/config.py PLUGIN_DIR resolution
- Updated all test imports (9 files)
- All 149 tests passing
