+++ __init__.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 import logging
 import threading
 
-from . import config, facts, ingester
+from cronalytics import config, facts, ingester
 
 
 def register(ctx) -> None:
@@ -24,7 +24,7 @@ def register(ctx) -> None:
     # gateway was down or the plugin was disabled.
     def _bootstrap():
         try:
-            from . import scanner
+            from cronalytics import scanner
             result = scanner.run_sync(config.STATE_DB, config.FACT_DB, config.WATERMARK_FILE)
             if result.get("inserted") or result.get("skipped"):
                 logging.getLogger("cronalytics").info(
