+++ cronalytics/__init__.py
@@ -1,0 +1,1 @@
+"""Cronalytics — cron observability package."""
