+++ tests/test_ingester.py
@@ -5,18 +5,13 @@ Uses importlib with fake package context so relative imports resolve.
 
 from __future__ import annotations
 
-import contextlib
 import importlib.util
 import json
 import sys
-import tempfile
-import threading
-import time
 from pathlib import Path
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
 import pytest
-
 
 # ---------------------------------------------------------------------------
 # Load ingester with fake package context (relative imports)
@@ -28,7 +23,7 @@ _PKG_NAME = "cronalytics_test_pkg"
 
 def _load_module(name: str):
     """Load a sibling module from the plugin root into the fake package."""
-    mod_path = _PLUGIN_ROOT / f"{name}.py"
+    mod_path = _PLUGIN_ROOT / "cronalytics" / f"{name}.py"
     full_name = f"{_PKG_NAME}.{name}"
     spec = importlib.util.spec_from_file_location(full_name, mod_path)
     mod = importlib.util.module_from_spec(spec)
@@ -49,7 +44,7 @@ for _dep in ("facts", "config", "logger"):
 
 # Load ingester itself
 _ingester_spec = importlib.util.spec_from_file_location(
-    f"{_PKG_NAME}.ingester", _PLUGIN_ROOT / "ingester.py"
+    f"{_PKG_NAME}.ingester", _PLUGIN_ROOT / "cronalytics" / "ingester.py"
 )
 ingester = importlib.util.module_from_spec(_ingester_spec)
 ingester.__package__ = _PKG_NAME
