+++ tests/test_scanner.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import json
 from pathlib import Path
 
-import scanner
+from cronalytics import scanner
 
 
 class TestWatermark:
