+++ tests/test_scanner_run_sync.py
@@ -6,8 +6,7 @@ import json
 from pathlib import Path
 from unittest.mock import patch
 
-import facts
-import scanner
+from cronalytics import facts, scanner
 
 
 class TestRunSync:
